+++ src/kodo/elimination_decoder.hpp
@@ -1,0 +1,74 @@
+// Copyright Steinwurf ApS 2011-2013
+// Distributed under the "STEINWURF RESEARCH LICENSE 1.0".
+// See accompanying file LICENSE.rst or
+// http://www.steinwurf.com/licensing
+
+#pragma once
+
+#include <fifi/default_field.hpp>
+
+#include "payload_decoder.hpp"
+#include "symbol_id_decoder.hpp"
+#include "plain_symbol_id_reader.hpp"
+#include "aligned_coefficients_decoder.hpp"
+#include "forward_linear_block_decoder.hpp"
+#include "symbol_decoding_status_tracker.hpp"
+#include "symbol_decoding_status_counter.hpp"
+#include "coefficient_storage.hpp"
+#include "deep_symbol_storage.hpp"
+#include "storage_bytes_used.hpp"
+#include "storage_block_info.hpp"
+#include "final_coder_factory_pool.hpp"
+#include "finite_field_math.hpp"
+#include "finite_field_info.hpp"
+#include "coefficient_info.hpp"
+#include "elimination_coefficient_info.hpp"
+#include "elimination_coefficient_value_access.hpp"
+#include "elimination_coefficient_offset.hpp"
+
+namespace kodo
+{
+
+    /// @ingroup fec_stacks
+    ///
+    /// @brief The elimination decoder uses a non-square decoding
+    ///        matrix to ignore certain coding coding
+    ///        coefficients. One typical use for this is create an
+    ///        elimination decoer for a specific part of encoded
+    ///        symbols when the elimination decoder reaches full rank
+    ///        it will effectively be able to eliminate that part for
+    ///        the encoding, while still producing valid encoded
+    ///        symbols.
+    template<class Field>
+    class elimination_decoder :
+        public // Decoder API
+               forward_linear_block_decoder<
+               symbol_decoding_status_counter<
+               symbol_decoding_status_tracker<
+               // Coefficient Storage API
+               coefficient_storage<
+               elimination_coefficient_value_access<
+               elimination_coefficient_info<
+               elimination_coefficient_offset<
+               coefficient_value_access<
+               coefficient_info<
+               // Storage API
+               deep_symbol_storage<
+               storage_bytes_used<
+               storage_block_info<
+               // Finite Field API
+               finite_field_math<typename fifi::default_field<Field>::type,
+               finite_field_info<Field,
+               // Factory API
+               final_coder_factory_pool<
+               // Final type
+               elimination_decoder<Field>
+               > > > > > > > > > > > > > > >
+    { };
+
+}
+
+
+
+
+